+++ src/main/java/store/lastdance/controller/calendar/CalendarV2Controller.java
@@ -15,9 +15,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import store.lastdance.dto.calender.request.CreateCalendarRequestDTO;
-import store.lastdance.dto.calender.request.UpdateCalendarRequestDTO;
-import store.lastdance.dto.calender.response.CalendarResponseDTO;
+import store.lastdance.dto.calendar.request.CreateCalendarRequestDTO;
+import store.lastdance.dto.calendar.request.UpdateCalendarRequestDTO;
+import store.lastdance.dto.calendar.response.CalendarResponseDTO;
 import store.lastdance.dto.common.ErrorResponseDTO;
 import store.lastdance.dto.response.ApiResponse;
 import store.lastdance.exception.CustomException;
