+++ src/main/java/store/lastdance/dto/calendar/request/CreateCalendarRequestDTO.java
@@ -1,4 +1,4 @@
-package store.lastdance.dto.calender.request;
+package store.lastdance.dto.calendar.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,18 +6,20 @@ import lombok.NoArgsConstructor;
 import store.lastdance.domain.calendar.CalendarCategory;
 import store.lastdance.domain.calendar.RepeatType;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
-import java.util.UUID;
 
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UpdateCalendarRequestDTO {
+public class CreateCalendarRequestDTO {
 
+    @NotBlank(message = "제목은 필수입니다.")
     @Size(max = 200, message = "제목은 200자 이내로 입력해주세요.")
     private String title;
 
@@ -28,13 +30,13 @@ public class UpdateCalendarRequestDTO {
 
     private LocalDateTime endDate;
 
+    @NotNull(message = "종일 여부는 필수입니다.")
     private Boolean isAllDay;
 
+    @NotNull(message = "카테고리는 필수입니다.")
     private CalendarCategory category;
 
-    private UUID groupId;
-
-    private RepeatType repeatType;
+    private RepeatType repeatType = RepeatType.NONE;
 
     private LocalDateTime repeatEndDate;
 }
