+++ src/test/java/store/lastdance/service/calendar/CalendarServiceTest.java
@@ -15,9 +15,9 @@ import store.lastdance.domain.group.Group;
 import store.lastdance.domain.user.OAuthProvider;
 import store.lastdance.domain.user.User;
 import store.lastdance.domain.user.UserRole;
-import store.lastdance.dto.calender.request.CreateCalendarRequestDTO;
-import store.lastdance.dto.calender.request.UpdateCalendarRequestDTO;
-import store.lastdance.dto.calender.response.CalendarResponseDTO;
+import store.lastdance.dto.calendar.request.CreateCalendarRequestDTO;
+import store.lastdance.dto.calendar.request.UpdateCalendarRequestDTO;
+import store.lastdance.dto.calendar.response.CalendarResponseDTO;
 import store.lastdance.exception.CustomException;
 import store.lastdance.exception.ErrorCode;
 import store.lastdance.repository.calendar.CalendarRepository;
